fix(lesson): sanitize lesson server side
